Set Up the translation package and restaurants ownership checking for preventing unauthorized actions.

+++ app/Http/Controllers/Api/CategoriesController.php
@@ -4,9 +4,11 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
+use Illuminate\Validation\Rule;
 
 class CategoriesController extends Controller
 {
@@ -15,15 +17,14 @@ class CategoriesController extends Controller
      */
     public function index(Request $request)
     {
-        $restaurant = $request->user()->restaurant;
+        $category = Category::where('restaurant_id', $request->header('X-Restaurant-Id') ?? $request->input('restaurant_id')
+            ?? $request->route('restaurant')
+            ?? $request->route('id'))->get();
 
-        $restCates = $restaurant->categories()->orderBy('name')->get();
-        if (count($restCates) >= 0) {
-            return response()->json(
-                response()->json([
-                    'data' => $restCates,
-                ])
-            );
+        if (!empty($category)) {
+            return response()->json([
+                    'data' => $category,
+            ]);
         } else {
             return response()->json([
                 'Message' => 'No Categories for this restaurant!',
@@ -44,13 +45,28 @@ class CategoriesController extends Controller
      */
     public function store(Request $request)
     {
-        $restaurant = $request->user()->restaurant;
+        // $user = auth()->user();
+
+        // $authorized = $user->restaurants()
+        //     ->where('restaurants.id', $request->restaurant_id)
+        //     ->exists();
+
+        // if (!$authorized) {
+        //     return response()->json([
+        //         'message' => 'Unauthorized restaurant'
+        //     ], 403);
+        // }
 
         $validated = $request->validate([
-            'name'        => 'required|string|max:255',
-            'description' => 'nullable|string',
-            'image'       => 'nullable|image|mimes:jpg,png,jpeg,webp|max:2048',
-            'is_active'   => 'boolean',
+            'restaurant_id'     => 'required|exists:restaurants,id',
+            'name'              => 'required|array',
+            'name.ar'           => ['required', 'string', 'max:255', Rule::unique('categories', 'name->ar')],
+            'name.en'           => ['required', 'string', 'max:255', Rule::unique('categories', 'name->en')],
+            'description'       => 'nullable|array',
+            'description.ar'    => 'nullable|string|min:3|max:1000',
+            'description.en'    => 'nullable|string|min:3|max:1000',
+            'image'             => 'nullable|image|mimes:jpg,png,jpeg,webp|max:2048',
+            'is_active'         => 'boolean',
         ]);
 
         if ($request->hasFile('image') && $request->file('image')->isValid()) {
@@ -60,7 +76,9 @@ class CategoriesController extends Controller
             $filename = Str::uuid() . '.' . $file->getClientOriginalExtension();
 
             $validated['image'] = $file->storeAs(
-                'categories', $filename, 'public'
+                'categories',
+                $filename,
+                'public'
             );
         }
 
@@ -99,13 +117,18 @@ class CategoriesController extends Controller
      */
     public function update(Request $request, Category $category)
     {
-        $restaurant = $request->user()->restaurant;
+        $restaurant = auth()->user()->restaurant;
 
         $validated = $request->validate([
-            'name'        => 'sometimes|string|max:255',
-            'description' => 'sometimes|nullable|string',
-            'image'       => 'sometimes|nullable|image|mimes:jpg,png,jpeg,webp|max:2048',
-            'is_active'   => 'boolean',
+            'restaurant_id'     => 'required|exists:restaurants,id',
+            'name'              => 'required|array',
+            'name.ar'           => ['required', 'string', 'max:255', Rule::unique('categories', 'name->ar')],
+            'name.en'           => ['required', 'string', 'max:255', Rule::unique('categories', 'name->en')],
+            'description'       => 'nullable|array',
+            'description.ar'    => 'nullable|string|min:3|max:1000',
+            'description.en'    => 'nullable|string|min:3|max:1000',
+            'image'             => 'nullable|image|mimes:jpg,png,jpeg,webp|max:2048',
+            'is_active'         => 'boolean',
         ]);
 
         if ($request->hasFile('image') && $request->file('image')->isValid()) {
@@ -114,7 +137,7 @@ class CategoriesController extends Controller
             }
 
             $file = $request->file('image');
-            $filename = $category->name . rand(1,9999) . '.' . $file->getClientOriginalExtension();
+            $filename = Carbon::now()->timestamp . rand(1, 9999) . '.' . $file->getClientOriginalExtension();
             $path = $file->storeAs('categories', $filename, 'public');
 
             $validated['image'] = $path;
@@ -133,8 +156,25 @@ class CategoriesController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Category $category)
+    public function destroy(Request $request, Category $category)
     {
+        $user = auth()->user();
+        $restaurantId = $request->header('X-Restaurant-Id') ?? $request->input('restaurant_id')
+            ?? $request->route('restaurant')
+            ?? $request->route('id');
+
+        $restaurant = $user->restaurants()->where('restaurants.id', $restaurantId)->first();
+
+         if ($category->restaurant_id !== $restaurant->id) {
+            return response()->json([
+                'message' => 'Unauthorized category'
+            ], 403);
+        }
+
+         if ($category->image) {
+            Storage::disk('public')->delete($category->image);
+        }
+
         $category->delete();
         return response()->json([
             'Message' => 'Category Deleted Successfully.',
